+++ docs/Changelog.md
@@ -2,6 +2,7 @@
 
 ## New since 0.10.0
 
+- PathLayer(s) scaled width [#594](https://github.com/mapsforge/vtm/issues/594)
 - vtm-models module [#580](https://github.com/mapsforge/vtm/issues/580)
 - Many other minor improvements and bug fixes
 - [Solved issues](https://github.com/mapsforge/vtm/issues?q=is%3Aclosed+milestone%3A0.11.0)